+++ JotTouchSDK.framework/Versions/A/Headers/JotPalmGestureRecognizer.h
@@ -1,0 +1,12 @@
+//
+//  JotPalmGestureRecognizer.h
+//  JotTouchSDK
+//
+//  Created on 1/29/13.
+//  Copyright (c) 2013 Adonit. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface JotPalmGestureRecognizer : UIGestureRecognizer
+@end