+++ JotTouchSDK.framework/Versions/A/Headers/JotTouchSDK.h
@@ -1,0 +1,18 @@
+//
+//  JotTouchSDK.h
+//  JotTouchSDK
+//
+//  Created by Timothy Ritchey on 4/22/14.
+//  Copyright (c) 2014 Adonit. All rights reserved.
+//
+
+#ifndef JotTouchSDK_JotTouchSDK_h
+#define JotTouchSDK_JotTouchSDK_h
+
+#import <JotTouchSDK/JotStylusManager.h>
+#import <JotTouchSDK/JotTapToConnectSettingsViewController.h>
+#import <JotTouchSDK/JotStylusMotionManager.h>
+#import <JotTouchSDK/JotPalmGestureRecognizer.h>
+#import <JotTouchSDK/UITouch+JotStylus.h>
+
+#endif