+++ JotTouchSDK.framework/Versions/A/Headers/UITouch+JotStylus.h
@@ -1,0 +1,13 @@
+//
+//  UITouch+JotStylus.h
+//  JotTouchSDK
+//
+//  Created by Timothy Ritchey on 3/3/14.
+//  Copyright (c) 2014 Adonit. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface UITouch (JotStylus)
+@property (assign, nonatomic) BOOL isJotStylus;
+@end